gdm-display: Add gdm_display_schedule_kill_greeter method
Once a user has logged in, the greeter session is no longer needed until
the user logs out; or user-switching is used.

The greeter-session takes quite a large amount of RAM, as such it
is undesirable to keep the greeter-session running all the time.

This commit adds a new gdm_display_schedule_kill_greeter method which can
be used to schedule stopping the greeter-session, and un-managing the
greeter display, X seconds after calling it.

Also add a counterpart gdm_display_cancel_schedule_kill_greeter method,
which can be called to keep te greeter around if the user-session exits
before the timeout expires.