local-display-factory: In seat0 create display if there is no greeter
Now that we stop the greeter session when it is no longer necessary,
the following can happen:
1) Log in as user a.
2) User-switch to user b.
3) Log out user b.
4) The logout takes you back to VT1, but there is no greeter session
running.

The problem here is that the create_display() call from
gdm_local_display_factory_sync_seats() does not create a new display
because there already is a display on the seat for user a.

This commit fixes this by skipping non-greeter displays when looking
for an existing display on seat0.